Handling http favicon bug
There's an issue with the web application that seems to push an http version of the favicon when it should be https. Tried a lot of suggestions from nopCommerce and continued to have issues. Even saw that there used to be a feature that handled this bug (Force SSL ) that was removed in a prior release. Decided to just handle it at the Application Layer.

A ToDo would be to handle this through the CMS Admin Backend and allow for an option as a bypass that detects it. Haven't spent enough time learning the way the application is structured to implement that fix. A more ideal approach would be to see what causes this issue.